BUSS6002 Assignment 1&2
Assignment 1
The Australian federal government is building a website to provide households with a tool to
determine if installing a rooftop solar panel system is right for them. On the website, users will
be able to enter information about their house and the website will provide an estimate of the
possible solar power generation.
The government collected a random sample of existing households with solar panels, including information about the households, solar panel installation and the associated solar power
generation. The generation from the solar panels was collected from 1/1/2022 to 31/12/2022.
The data has been assembled from a multiple sources including the customer energy retailer,
energy distributors and solar installers. Sampling is limited to installations with:
• a single solar panel array or multiple arrays that are oriented identically,
• rooftop installation only.
As a data-scientist-in-training, you will assist the project by completing a variety of EDA
tasks.

Assignment 2
The Australian federal government is building a website to provide households with a tool to
determine if installing a rooftop solar panel system is right for them. On the website, users will
be able to enter information about their house and the website will provide an estimate of the
possible solar power generation.
The government collected a random sample of existing households with solar panels, including information about the households, solar panel installation and the associated solar power
generation. The generation from the solar panels was collected from 1/1/2022 to 31/12/2022.
The data has been assembled from a multiple sources including the customer energy retailer,
energy distributors and solar installers. Sampling is limited to installations with:
• a single solar panel array or multiple arrays that are oriented identically,
• rooftop installation only.
After you presented your EDA from Assignment 1, you have been given a new task: determine if a model can be built to predict Generation, that can outperform simple baselines.